Defining the contract that I need to implement.  Yes, this really was my first step including sticking in the template parameter.
I suppose if I had really been fair, I would have used Probability as the name for
my implementation, and then done the Extract Interface refactoring when I hit the
point where I needed to swap out the implementation.  But I've been thinking a lot
about integrated tests lately, so this just fell out.

Yes, with the spelling errors in the return values of the method signatures.
Because I'm an idiot.

I'm not conforming to Greg's method naming conventions, because he's not the boss of me.