RED: introducing a first test.
Continuing to cheat with integrated test patterns in mind.  Again,
I'm building a "suite" that can be used with multiple implementations
of the contract.  Using a data provider will also give me an easy way
to check my tests against multiple values without a lot of extra effort.

This test came from staring at the contract for a long time, trying
to figure out where the seam to test the state is.  Missing a key
insight, I was starting from the idea of "given a Probability, where
do I get another one?"  Since there's only one generator in the
interface, we "have" to start there.

WRONG.

But let's go with it.  My dumbitude will be exposed in due course.