GREEN: checkIdentity
It's about this point in the game that I realized that the first check
I needed was an equality check; I'll have to double back to follow
that line of thought later.  It bothered me for a while that I needed
to introduce equals on my own.  I'll find a loophole to allow it later.