RED Introduce a test for combineWith
I miss some more opportunities to explore inverseOf().  Instead, I look
for a way to work in the other two methods.  I finally realize that the
seam I have to work with is the symmetry of the probability rules

X(A,B) = X(B,A)

So that's where my next test needs to go.  I choose combineWith because
it's the simpler of the two operations.  I write a simple test, and it
passes!  OOPS.  I shim in a couple of null checks to get it to fail
as it should.