GREEN Disable the failing test.
I try to implement combineWith() in a straight forward way, and
discover that I screwed up the contract on the value object, such
that I can't get at the state of the other object.  In other words,
two _different_ implementations of the Probability contract shouldn't
be expected to play nice.

Actually, this reminds me of a blocking problem I had with Monads...
how the check do you get the value out afterwards?  Oh, said Master
Scala, the Monad has to implement something _other_ than the monad
interface to be useful....  Good to know.