GREEN Fixing the specification of the contract.
Bah - NOT a mistake I would have made if I had written a real implementation
first, and extracted the contract out of it afterwards.  I've used that
pattern often enough that I should know better; but my attention lapsed,
and I had to proceed a while to catch the error.