GREEN Refactoring to clarify the value tests
I realize in looking at the symmetry of operations, that there's
a bit of duplication in the organization of my checks.  Also, I
realized the value checks should take a value interface, it keeps
me from cheating, and allows me to re-use them later.

I also decided to physically move the checks around, so that they
were closer to the relevant interface.

It finally occurred to me that I might want to have some tests
with two arguments, since there are, after all, a number of
expressions with two different arguments in them.