GREEN Refactoring
I realized that having the functions ordered in the class definition
as I was choosing to test them would have prevented the previous mistake.

Also, I decided as a matter of aesthetics, it's more expressive to write
that we are creating the probability from a double, rather than just
using a constructor.