RED Broken compile!
There's something the symmetry of combinedWith that gets me looking
at the fact that I didn't really ensure that inverseOf has the
proper symmetry to it.  So I add that check.

It immediately exposes a bug in the same test where I made the change,
because I missed a spot when I was cleaning up my earlier mistake with
the generics.

Perhaps I should add another unit test to measure the probability that
I will run into this mistake again.