GREEN New check of inverseOf
I realized in looking at the previous implementation of this check
that I had made a mistake -- all my check actually did was make
sure that start.inverseOf() returned the same value twice in a row;
what I wanted to check was that not(not(X)) == X was also true of
not(X).

But now the symmetry is broken.