GREEN Refactored test to improve symmetry.
There's a pattern starting to show up in these tests, which is that
given arguments start and other, I always seem to have a check that
uses them in that order, and a check that uses them in the other
order.

For instance, in checkInverseOf, I first check start while ignoring
other, then I check other while ignoring start.  i

In other words, I'm always putting in an extra check to ensure I
haven't screwed up the fact that doubles commute under both
multiplication and addition.

That's duplication, but worse is that I have a growing suspicion
that it is going to bite me in the ass....