GREEN: Refactor: rename Method
Changed the spelling of the method to more accurately reflect its intent.

It is shaped, sir, like itself, and it is as broad as it hath breadth.
It is just so high as it is, and moves with its own organs. It lives
by that which nourisheth it, and, the elements once out of it, it transmigrates.

Of its own color too.