RED Restored the broken test case.
OK, now that I see what's going on, I see that I have an extra out.  So I'm
going to take it.

The problem here is that TestDouble uses floating point math, and floating
point math is not the same thing as real math.  In particular, when you are
working with numbers that can't be exactly represented, you run the risk
of some weird rounding errors.

Old lesson, when checking floating point numbers for equality, you usually
want to include some tolerance.