RED Test fails to compile.
Watching in slow motion, it's weird to me that one compile error
in the test indicates an invalid test, and the next compile error
in the test indicates an invalid implementation.

I suppose this is trying to show me that the discipline is to
keep writing test until something bricks.