GREEN What???
OK, the simplest thing that couple possibly work is to accept the
function signature suggested by my IDE.  Doing so gives me a passing
test with the wrong signature.

Why? because I made two errors; the first was thinking that introducing
the name of the method would suffice to "get it right", and also
because I did keep writing the test until it bricked.