GREEN Refactoring
Decided to run the method signature direction, as it gives me a concrete
thing that I can call "done" -- the business value, such as it is, is that
I can provide to other developers an implementation that they can use to
write components that integrate with mine.

It's still a bit fuzzy, but I don't feel like a complete fraud at this
point.