GREEN: First complete test is passing.
SHIP IT ! We've done just barely enough work that we can demonstate
with tests that we have unblocked any developers waiting to compile
against our implementation.

So we could put this in a jar file, with a specification that
promises nothing, and away we go.

So that's really the end of Greg's kata, as far as I'm concerned;
I've figured out what the first test to write is.

That said, I need to push on -- there's is local minimum that I need
to explore.