RED Failing test: NPE
This looks a bit weird, but if the negation of the negation is an
identity, then that ought to be true for any Probabilty we use
including those that are results of other computations.

Obviously, there is some duplication present, but I need to
press ahead to complete the duplication before I can fairly
address it.